# backend/app/schemas/ingrediente.py
from typing import Optional
from pydantic import BaseModel, ConfigDict

class IngredienteBase(BaseModel):
    nombre: str
    descripcion: Optional[str] = None
    es_alergeno: bool = False

class IngredienteCreate(IngredienteBase):
    pass

class IngredienteUpdate(BaseModel):
    nombre: Optional[str] = None
    descripcion: Optional[str] = None
    es_alergeno: Optional[bool] = None

class IngredienteRead(IngredienteBase):
    id: int
    
    model_config = ConfigDict(from_attributes=True)