# frontend/parcial1/openapi.json
{"openapi":"3.1.0","info":{"title":"API de Gestión de Productos y Categorías","description":"Sistema Backend para Parcial Universitario con FastAPI, SQLModel y PostgreSQL","version":"1.0.0"},"paths":{"/categorias/":{"get":{"tags":["Categorias"],"summary":"Listar todas las categorías","operationId":"listar_categorias_categorias__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoriaRead"},"type":"array","title":"Response Listar Categorias Categorias  Get"}}}}}},"post":{"tags":["Categorias"],"summary":"Crear una nueva categoría","operationId":"crear_categoria_categorias__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriaCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriaRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categorias/{id}":{"get":{"tags":["Categorias"],"summary":"Obtener categoría por ID","operationId":"obtener_categoria_categorias__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"El ID de la categoría"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriaRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Categorias"],"summary":"Actualizar una categoría","operationId":"actualizar_categoria_categorias__id__patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriaUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoriaRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Categorias"],"summary":"Eliminar una categoría","operationId":"eliminar_categoria_categorias__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/productos/":{"get":{"tags":["Productos"],"summary":"Listar todos los productos","description":"Retorna una lista de todos los productos registrados en el sistema.","operationId":"listar_productos_productos__get","parameters":[{"name":"disponible","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Disponible"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductoRead"},"title":"Response Listar Productos Productos  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Productos"],"summary":"Crear un nuevo producto","operationId":"crear_producto_productos__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/productos/{id}":{"get":{"tags":["Productos"],"summary":"Obtener producto por ID","operationId":"obtener_producto_productos__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Productos"],"summary":"Actualizar un producto existente","operationId":"actualizar_producto_productos__id__patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Productos"],"summary":"Eliminar un producto","operationId":"eliminar_producto_productos__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ingredientes/":{"get":{"tags":["Ingredientes"],"summary":"List Ingredientes","operationId":"list_ingredientes_ingredientes__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/IngredienteRead"},"type":"array","title":"Response List Ingredientes Ingredientes  Get"}}}}}},"post":{"tags":["Ingredientes"],"summary":"Create Ingrediente","operationId":"create_ingrediente_ingredientes__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredienteCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredienteRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ingredientes/{id}":{"get":{"tags":["Ingredientes"],"summary":"Get Ingrediente","operationId":"get_ingrediente_ingredientes__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredienteRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Ingredientes"],"summary":"Update Ingrediente","operationId":"update_ingrediente_ingredientes__id__patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredienteUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngredienteRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Ingredientes"],"summary":"Delete Ingrediente","operationId":"delete_ingrediente_ingredientes__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CategoriaCreate":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"imagen_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Imagen Url"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"}},"type":"object","required":["nombre"],"title":"CategoriaCreate"},"CategoriaRead":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"imagen_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Imagen Url"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["nombre","id"],"title":"CategoriaRead"},"CategoriaUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"imagen_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Imagen Url"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"}},"type":"object","title":"CategoriaUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngredienteCreate":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"es_alergeno":{"type":"boolean","title":"Es Alergeno","default":false}},"type":"object","required":["nombre"],"title":"IngredienteCreate"},"IngredienteRead":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"es_alergeno":{"type":"boolean","title":"Es Alergeno","default":false},"id":{"type":"integer","title":"Id"}},"type":"object","required":["nombre","id"],"title":"IngredienteRead"},"IngredienteUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"es_alergeno":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Es Alergeno"}},"type":"object","title":"IngredienteUpdate"},"ProductoCreate":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"type":"string","title":"Descripcion"},"precio_base":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Precio Base"},"imagenes_url":{"items":{"type":"string"},"type":"array","title":"Imagenes Url"},"stock_cantidad":{"type":"integer","title":"Stock Cantidad","default":0},"disponible":{"type":"boolean","title":"Disponible","default":true},"categoria_ids":{"items":{"type":"integer"},"type":"array","title":"Categoria Ids","default":[]},"ingrediente_ids":{"items":{"type":"integer"},"type":"array","title":"Ingrediente Ids","default":[]}},"type":"object","required":["nombre","descripcion","precio_base","imagenes_url"],"title":"ProductoCreate"},"ProductoRead":{"properties":{"nombre":{"type":"string","title":"Nombre"},"descripcion":{"type":"string","title":"Descripcion"},"precio_base":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Precio Base"},"imagenes_url":{"items":{"type":"string"},"type":"array","title":"Imagenes Url"},"stock_cantidad":{"type":"integer","title":"Stock Cantidad","default":0},"disponible":{"type":"boolean","title":"Disponible","default":true},"id":{"type":"integer","title":"Id"}},"type":"object","required":["nombre","descripcion","precio_base","imagenes_url","id"],"title":"ProductoRead"},"ProductoUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"precio_base":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Precio Base"},"imagenes_url":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Imagenes Url"},"stock_cantidad":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Stock Cantidad"},"disponible":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Disponible"},"categoria_ids":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Categoria Ids"},"ingrediente_ids":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Ingrediente Ids"}},"type":"object","title":"ProductoUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}